# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/recruit": {
            "post": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Recruit a user",
                "description": "Perusahaan mitra melakukan recruitment terhadap user.",
                "operationId": "recruit",
                "requestBody": {
                    "description": "ID user yang direcruit",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ID_user"
                                ],
                                "properties": {
                                    "ID_user": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model user"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/recruit/confirm/{id}": {
            "post": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Confirm recruitment",
                "description": "Perusahaan mitra mengonfirmasi status recruitment user.",
                "operationId": "confirmRecruitment",
                "parameters": [
                    {
                        "name": "ID_recruitment",
                        "in": "path",
                        "description": "ID recruitment yang akan dikonfirmasi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Status recruitment baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "tersedia",
                                            "menunggu",
                                            "tidak tersedia"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model recruitment"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/recruitments": {
            "get": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Show recruited users by logged in company partner",
                "description": "Perusahaan mitra melihat daftar user yang direkrut olehnya.",
                "operationId": "showRecruitment",
                "responses": {
                    "default": {
                        "description": "return array model recruitment"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}